Implement scoreboard
Implements the scoreboard as a singleton JPanel object. The scoreboard's refresh method is called when the square button click listener is triggered. A getter method for the list of players in the board model is also added for use in the scoreboard. The KeyCollector window size is adjusted to be wider to accommodate the scoreboard that is placed to the right of the board view.