Adjust clearBoard to only clear the square's data
Previously, clearBoard would overwrite the board's existing square instances with new ones. However, this would mess with Piece reconstruction from formatted strings, as the constructors relied on getting the square instances that are currently on the board to assign to the pieces. To avoid this problem, a new clear method is added to the Square class, which clears the square's player list and special piece attribute. The Board class's clearBoard method now calls this clear method instead of replacing the square instances.

This fixes an issue where pieces would not appear or duplicate pieces would appear when loading from a saved file.